Changed to f32 function

+++ src/vec2.rs
@@ -14,7 +14,7 @@ impl PartialEq<Vec2> for Vec2 {
 
 impl Vec2 {
     pub fn distance(v1: &Vec2, v2: &Vec2) -> f32 {
-        ((v2.x - v1.x).pow(2) as f32 + (v2.y - v1.y).pow(2) as f32).sqrt()
+        ((v2.x - v1.x).powf(2.0) as f32 + (v2.y - v1.y).powf(2.0) as f32).sqrt()
     }
 
     pub fn move_towards(&mut self, to: Vec2, step: f32) {
